// Atspm/WebUI/src/api/config/aTSPMConfigurationApi.schemas.ts
export interface Approach {
  /** @nullable */
  description?: string | null
  /** @nullable */
  mph?: number | null
  protectedPhaseNumber?: number
  isProtectedPhaseOverlap?: boolean
  /** @nullable */
  permissivePhaseNumber?: number | null
  isPermissivePhaseOverlap?: boolean
  /** @nullable */
  pedestrianPhaseNumber?: number | null
  isPedestrianPhaseOverlap?: boolean
  /** @nullable */
  pedestrianDetectors?: string | null
  locationId?: number
  location?: Location
  directionTypeId?: DirectionTypes
  directionType?: DirectionType
  /** @nullable */
  detectors?: Detector[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
}
// ...
export interface Area {
  /** @nullable */
  name?: string | null
  /** @nullable */
  userAreas?: UserArea[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}
// ...
export interface Device {
  /** @nullable */
  deviceIdentifier?: string | null
  loggingEnabled?: boolean
  /** @nullable */
  ipaddress?: string | null
  deviceStatus?: DeviceStatus
  deviceType?: DeviceTypes
  /** @nullable */
  notes?: string | null
  locationId?: number
  /** @nullable */
  deviceConfigurationId?: number | null
  deviceConfiguration?: DeviceConfiguration
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  location?: Location
}

export interface DeviceConfiguration {
  /** @nullable */
  description?: string | null
  /** @nullable */
  notes?: string | null
  protocol?: TransportProtocols
  port?: number
  /** @nullable */
  path?: string | null
  /** @nullable */
  query?: string[] | null
  connectionTimeout?: number
  operationTimeout?: number
  loggingOffset?: number
  /** @nullable */
  decoders?: string[] | null
  /** @nullable */
  userName?: string | null
  /** @nullable */
  password?: string | null
  /** @nullable */
  productId?: number | null
  product?: Product
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  devices?: Device[] | null
}

export interface DeviceGroup {
  /** @nullable */
  manufacturer?: string | null
  /** @nullable */
  model?: string | null
  /** @nullable */
  firmware?: string | null
  count?: number
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeviceStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const

export type DeviceTypes = (typeof DeviceTypes)[keyof typeof DeviceTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeviceTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
} as const
// ...
export interface Jurisdiction {
  /** @nullable */
  name?: string | null
  /** @nullable */
  mpo?: string | null
  /** @nullable */
  countyParish?: string | null
  /** @nullable */
  otherPartners?: string | null
  /** @nullable */
  userJurisdictions?: UserJurisdiction[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}
// ...
export interface Location {
  latitude?: number
  longitude?: number
  /** @nullable */
  primaryName?: string | null
  /** @nullable */
  secondaryName?: string | null
  chartEnabled?: boolean
  versionAction?: LocationVersionActions
  /** @nullable */
  note?: string | null
  start?: string
  pedsAre1to1?: boolean
  /** @nullable */
  locationIdentifier?: string | null
  /** @nullable */
  jurisdictionId?: number | null
  jurisdiction?: Jurisdiction
  locationTypeId?: number
  locationType?: LocationType
  /** @nullable */
  regionId?: number | null
  region?: Region
  /** @nullable */
  areas?: Area[] | null
  /** @nullable */
  devices?: Device[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  approaches?: Approach[] | null
}
// ...
export interface LocationType {
  /** @nullable */
  name?: string | null
  /** @nullable */
  icon?: string | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}

export type LocationVersionActions =
  (typeof LocationVersionActions)[keyof typeof LocationVersionActions]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LocationVersionActions = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_10: 10,
} as const
// ...
export interface Product {
  /** @nullable */
  manufacturer?: string | null
  /** @nullable */
  model?: string | null
  /** @nullable */
  webPage?: string | null
  /** @nullable */
  notes?: string | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
}
// ...
export interface Region {
  /** @nullable */
  description?: string | null
  /** @nullable */
  userRegions?: UserRegion[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}
// ...
export type TransportProtocols =
  (typeof TransportProtocols)[keyof typeof TransportProtocols]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransportProtocols = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
} as const